[CLEANUP] Full Lint Compliance

See merge request thorchain/thornode!2316

+++ bifrost/pkg/chainclients/binance/binance.go
@@ -602,7 +602,7 @@ func (b *Binance) ReportSolvency(bnbBlockHeight int64) error {
 			b.logger.Err(err).Msgf("fail to get account balance")
 			continue
 		}
-		if runners.IsVaultSolvent(acct, asgard, cosmos.NewUint(3*uint64(b.bnbScanner.singleFee))) && b.IsBlockScannerHealthy() {
+		if runners.IsVaultSolvent(acct, asgard, cosmos.NewUint(3*b.bnbScanner.singleFee)) && b.IsBlockScannerHealthy() {
 			// when vault is solvent , don't need to report solvency
 			continue
 		}
